requirements.txt
yfinance>=0.2.66
python-dotenv>=1.2.1
openai>=2.6.0
fastapi>=0.120.3
pydantic>=2.12.3
uvicorn>=0.31.1
httpx>=0.27.0